ai-agents/requirements.txt
dstack-sdk==0.1.2
asyncio>=3.4.3
fastapi==0.111.0
openai==v1.52.0
python-dotenv==1.0.0